+++ api/src/main/java/org/grnet/cat/api/endpoints/registry/RegistryCodelistEndpoint.java
@@ -27,6 +27,9 @@ import org.grnet.cat.dtos.registry.codelist.TypeCriterionResponse;
 import org.grnet.cat.repositories.registry.ImperativeRepository;
 import org.grnet.cat.repositories.registry.TypeCriterionRepository;
 import org.grnet.cat.services.registry.ImperativeService;
+import org.grnet.cat.dtos.registry.codelist.TypeBenchmarkResponse;
+import org.grnet.cat.repositories.registry.TypeBenchmarkRepository;
+import org.grnet.cat.services.registry.TypeBenchmarkService;
 import org.grnet.cat.services.registry.TypeCriterionService;
 import org.grnet.cat.utils.Utility;
 
@@ -43,6 +46,8 @@ public class RegistryCodelistEndpoint {
     TypeCriterionService typeCriterionService;
     @Inject
     ImperativeService imperativeService;
+    @Inject
+    TypeBenchmarkService typeBenchmarkService;
 
     @ConfigProperty(name = "api.server.url")
     String serverUrl;
@@ -94,8 +99,8 @@ public class RegistryCodelistEndpoint {
             required = true,
             example = "pid_graph:3E109BBA",
             schema = @Schema(type = SchemaType.STRING))
-                                  @PathParam("id")
-                                  @Valid @NotFoundEntity(repository = TypeCriterionRepository.class, message = "There is no Type Criterion with the following id:") String id) {
+                                     @PathParam("id")
+                                     @Valid @NotFoundEntity(repository = TypeCriterionRepository.class, message = "There is no Type Criterion with the following id:") String id) {
 
         var typeCriterion = typeCriterionService.getTypeCriterionById(id);
         return Response.ok().entity(typeCriterion).build();
@@ -137,10 +142,10 @@ public class RegistryCodelistEndpoint {
     @Path("/type-criterion")
     public Response getTypeCriterionList(@Parameter(name = "page", in = QUERY,
             description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
-                                   @Parameter(name = "size", in = QUERY,
-                                           description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
-                                   @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
-                                   @Context UriInfo uriInfo) {
+                                         @Parameter(name = "size", in = QUERY,
+                                                 description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                         @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                         @Context UriInfo uriInfo) {
 
         var typeCriterionList = typeCriterionService.getTypeCriterionListByPage(page - 1, size, uriInfo);
 
@@ -202,12 +207,12 @@ public class RegistryCodelistEndpoint {
     @Produces(MediaType.APPLICATION_JSON)
     @Registration
     public Response getImpertive(@Parameter(
-            description = "The ID of the Type Criterion to retrieve.",
+            description = "The ID of the Imperative to retrieve.",
             required = true,
             example = "pid_graph:3E109BBA",
             schema = @Schema(type = SchemaType.STRING))
-                                     @PathParam("id")
-                                     @Valid @NotFoundEntity(repository = ImperativeRepository.class, message = "There is no Imperative with the following id:") String id) {
+                                 @PathParam("id")
+                                 @Valid @NotFoundEntity(repository = ImperativeRepository.class, message = "There is no Imperative with the following id:") String id) {
 
         var imperative = imperativeService.getImperativeById(id);
         return Response.ok().entity(imperative).build();
@@ -249,10 +254,10 @@ public class RegistryCodelistEndpoint {
     @Path("/imperative")
     public Response getImperativeList(@Parameter(name = "page", in = QUERY,
             description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
-                                         @Parameter(name = "size", in = QUERY,
-                                                 description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
-                                         @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
-                                         @Context UriInfo uriInfo) {
+                                      @Parameter(name = "size", in = QUERY,
+                                              description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                      @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                      @Context UriInfo uriInfo) {
 
         var imperativeList = imperativeService.getImperativeListByPage(page - 1, size, uriInfo);
 
@@ -270,6 +275,120 @@ public class RegistryCodelistEndpoint {
 
         @Override
         public void setContent(List<ImperativeResponse> content) {
+            this.content = content;
+        }
+    }
+
+
+    @Tag(name = "Codelist")
+    @Operation(
+            summary = "Get specific TypeBenchmark.",
+            description = "Returns a specific TypeBenchmark.")
+    @APIResponse(
+            responseCode = "200",
+            description = "The corresponding TypeBenchmark.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = ImperativeResponse.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "User has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "Not permitted.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Entity Not Found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+    @GET
+    @Path("/type-benchmark/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Registration
+    public Response getTypeBenchmark(@Parameter(
+            description = "The ID of the Type Benchmark to retrieve.",
+            required = true,
+            example = "pid_graph:3E109BBA",
+            schema = @Schema(type = SchemaType.STRING))
+                                     @PathParam("id")
+                                     @Valid @NotFoundEntity(repository = TypeBenchmarkRepository.class, message = "There is no TypeBenchmark with the following id:") String
+                                             id) {
+
+        var typeBenchmark = typeBenchmarkService.getTypeBenchmarkById(id);
+        return Response.ok().entity(typeBenchmark).build();
+    }
+
+    @Tag(name = "Codelist")
+    @Operation(
+            summary = "Get list of TypeBenchmark.",
+            description = "This endpoint retrieves all TypeBenchmarks." +
+                    "By default, the first page of 10 TypeBenchmark will be returned. You can tune the default values by using the query parameters page and size.")
+    @APIResponse(
+            responseCode = "200",
+            description = "List of TypeBenchmark.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = PageableTypeBenchmarkResponse.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "User has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "Not permitted.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    @Registration
+    @Path("/type-benchmark")
+    public Response getTypeBenchmarkList(@Parameter(name = "page", in = QUERY,
+            description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
+                                         @Parameter(name = "size", in = QUERY,
+                                                 description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                         @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                         @Context UriInfo uriInfo) {
+
+        var typeBenchmarkList = typeBenchmarkService.getTypeBenchmarkListByPage(page - 1, size, uriInfo);
+
+        return Response.ok().entity(typeBenchmarkList).build();
+    }
+
+    public static class PageableTypeBenchmarkResponse extends PageResource<TypeBenchmarkResponse> {
+
+        private List<TypeBenchmarkResponse> content;
+
+        @Override
+        public List<TypeBenchmarkResponse> getContent() {
+            return content;
+        }
+
+        @Override
+        public void setContent(List<TypeBenchmarkResponse> content) {
             this.content = content;
         }
     }